requirements.txt
streamlit~=1.32.0
pandas~=2.2.2
ultralytics~=8.3.103
numpy~=1.26.4
pillow~=10.3.0
scikit-learn~=1.4.2
tqdm~=4.66.4